Make backend run, migrate to python3.12
